fix: add soft delete, use form validator

+++ app/Http/Controllers/UserController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UserChangePasswordRequest;
+use App\Http\Requests\UserStoreRequest;
+use App\Http\Requests\UserUpdateRequest;
 use App\Services\UserService;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Validator;
 use App\Traits\APIResponse;
 
 class UserController extends ControllerWithGuard
@@ -15,7 +17,7 @@ class UserController extends ControllerWithGuard
 
     public function __construct(UserService $userService)
     {
-        parent::__construct(); 
+        parent::__construct();
         $this->userService = $userService;
     }
 
@@ -25,23 +27,9 @@ class UserController extends ControllerWithGuard
         return $this->responseSuccessWithData($result);
     }
 
-    public function store(Request $request)
+    public function store(UserStoreRequest  $request)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|unique:users',
-            'password' => 'required|string|min:6',
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json([
-                'success' => false,
-                'errors' => $validator->errors()
-            ], 422);
-        }
-
-        $result = $this->userService->store($request->name, $request->email, $request->password);
-
+        $result = $this->userService->store($request->full_name, $request->email, $request->password);
         return $this->responseSuccessWithData($result, true);
     }
 
@@ -50,47 +38,93 @@ class UserController extends ControllerWithGuard
         $user = $this->userService->show($id);
 
         if (!$user) {
-            return response()->json(['message' => 'User not found'], 404);
+            return $this->responseError('User not found', 404);
         }
 
         return $this->responseSuccessWithData($user);
     }
 
-    public function update($id, Request $request)
+    public function update($id, UserUpdateRequest $request)
     {
-        $result = $this->userService->update($id, ['name'=>($request->name), 'email' => ($request->email), 'password'=> ($request->password)]);
+        $user = $this->userService->show($id);
+
+        if (!$user) {
+            return $this->responseError('User not found', 404);
+        }
+
+        $result = $this->userService->update($id, [
+            'full_name' => ($request->full_name),
+            'email' => ($request->email),
+            'password' => ($request->password)
+        ]);
         return $this->responseSuccessWithData($result);
     }
 
+    /**
+     * Soft deletes a user by their ID.
+     *
+     * @param int $id The ID of the user to be soft deleted.
+     * @return \Illuminate\Http\JsonResponse A success response indicating the user has been soft deleted.
+     */
     public function destroy($id)
     {
+        $user = $this->userService->show($id);
+
+        if (!$user) {
+            return $this->responseError('User not found', 404);
+        }
+
         $this->userService->destroy($id);
         return $this->responseSuccess('User soft deleted successfully!');
     }
 
+    /**
+     * Get all the deleted users.
+     * @return \Illuminate\Http\JsonResponse A success response with data of the deleted users.
+     */
     public function getDeletedUsers()
     {
         $result = $this->userService->getDeletedUsers();
         return $this->responseSuccessWithData($result);
     }
 
+    /**
+     * Restore a deleted user by their ID.
+     *
+     * @param int $id The ID of the deleted user.
+     * @return \Illuminate\Http\JsonResponse A success response indicating the user has been restored.
+     */
     public function restore($id)
     {
         $this->userService->restore($id);
         return $this->responseSuccess('Deleted user restored successfully!');
     }
 
+    /**
+     * Delete user permanently by their ID.
+     *
+     * @param int $id The ID of the user.
+     * @return \Illuminate\Http\JsonResponse A success response indicating the user has been delted permanently.
+     */
     public function forceDelete($id)
     {
+        $user = $this->userService->show($id);
+
+        if (!$user) {
+            return $this->responseError('User not found', 404);
+        }
+
         $this->userService->forceDelete($id);
         return $this->responseSuccess('User permanently deleted!');
     }
 
-    public function changePassword($id, Request $request)
+    public function changePassword($id, UserChangePasswordRequest $request)
     {
-        $request->validate([
-            'password' => 'required|min:8|confirmed', 
-        ]);
+        $user = $this->userService->show($id);
+
+        if (!$user) {
+            return $this->responseError('User not found', 404);
+        }
 
         $result = $this->userService->changePassword($id, $request->password);
         return $this->responseSuccessWithData($result);
